+++ test/integration/utt_example_plugin/utt/plugins/report_view.py
@@ -29,17 +29,19 @@ class MyReportView(_v1.ReportView):
     def render(self, output: _v1.Output) -> None:
         _v1.SummaryView(self._report_model.summary_model).render(output)
 
-        if self._report_model.per_day:
+        if self._report_model.args.show_per_day:
             _v1.PerDayView(self._report_model.per_day_model).render(output)
         else:
             _v1.ProjectsView(self._report_model.projects_model).render(output)
 
         _v1.ActivitiesView(self._report_model.activities_model).render(output)
 
-        if (self._report_model.start_date == self._report_model.end_date) or self._report_model.show_details:
-            _v1.DetailsView(self._report_model.details_model, show_comments=self._report_model.show_comments).render(
-                output
-            )
+        if (
+            self._report_model.args.range.start == self._report_model.args.range.end
+        ) or self._report_model.args.show_details:
+            _v1.DetailsView(
+                self._report_model.details_model, show_comments=self._report_model.args.show_comments
+            ).render(output)
 
         MySubView(self._my_sub_model).render(output)